fix: throw not found exception if member does not exist

+++ src/main/java/com/compartetutiempo/timebank/member/MemberService.java
@@ -94,6 +94,9 @@ public class MemberService {
 
     @Transactional
     public void delete(Integer memberId) {
+        if (!memberRepository.existsById(memberId)) {
+            throw new ResourceNotFoundException("Member", "id", memberId);
+        }
         memberRepository.deleteById(memberId);
     }
